Move pytest fixture to conftest.py
