+++ tests/test_rand.py
@@ -13,7 +13,7 @@ import pytest
 
 from OpenSSL import rand
 
-from .util import NON_ASCII, tmpfile
+from .util import NON_ASCII
 
 
 class TestRand(object):
